Use factories to create top level speific generic event modules
